Write to SQS instead of S3
S3 is very slow to read when dealing with thousands of very small files.
Moreover, if we want to use S3 again a better way would be to read from
the SQS queue and make tar packages of ~1000 files, which can then be
placed in S3 and dealt with more efficiently.